+++ lib/simple_credomatic_paycon/configuration.rb
@@ -1,0 +1,10 @@
+module SimpleCredomaticPaycon
+  class Configuration
+    attr_accessor :show_trace
+
+    def initialize
+      @show_trace = false
+    end
+
+  end
+end