+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.1.0
+
+* Basic full implementation of Credomatic Paycon
+* Support for trace responses via configurations
+* Basic documentation
+
 # 0.0.1
 
 * Nothing important yet